requirements.txt
asn1crypto==0.24.0
asyncpg==0.18.3
attrs==19.1.0
backcall==0.1.0
cassandra-driver==3.19.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
conda==4.7.11
conda-package-handling==1.3.11
cryptography==2.7
csiphash==0.0.5
cycler==0.10.0
Cython==0.29.14
decorator==4.4.0
embed==0.0.0
et-xmlfile==1.0.1
googledrivedownloader==0.4
h5py==2.9.0
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
isodate==0.6.0
jdcal==1.4.1
jedi==0.15.1
joblib==0.13.2
kiwisolver==1.1.0
libarchive-c==2.8
matplotlib==3.1.1
mkl-fft==1.0.14
mkl-random==1.0.2
mkl-service==2.0.2
networkx==2.3
numpy==1.16.4
olefile==0.46
openpyxl==3.0.0
pandas==0.25.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
plyfile==0.7
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.2
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.2
rdflib==4.2.2
redis==3.3.11
requests==2.22.0
ruamel-yaml==0.15.46
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
torch==1.2.0
torch-cluster==1.4.4
torch-geometric==1.4.3
torch-scatter==1.3.1
torch-sparse==0.4.0
torch-spline-conv==1.1.0
torchbiggraph==1.dev1
torchvision==0.4.0a0+6b959ee
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
